Only Accept IP Sockets
While it's possible to handle ping requests, STUN only makes sense with
an IP source address. Even WireGuard cookies used for rate limiting are
defined to be calculated with the source IP address and port.
